# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Castor Orçamento API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /financeiro/categorias-despesa:
    get:
      operationId: listCategoriasDespesa
      tags: [financeiro]
      summary: Listar categorias de despesa disponíveis
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /financeiro/despesas:
    get:
      operationId: listDespesas
      tags: [financeiro]
      summary: Listar despesas do mês
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: mes
          in: query
          schema:
            type: number
        - name: ano
          in: query
          schema:
            type: number
        - name: categoria
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de despesas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Despesa"
    post:
      operationId: criarDespesa
      tags: [financeiro]
      summary: Criar despesa
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarDespesaInput"
      responses:
        "200":
          description: Despesa criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
  /financeiro/despesas/{id}:
    put:
      operationId: updateDespesa
      tags: [financeiro]
      summary: Atualizar despesa
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarDespesaInput"
      responses:
        "200":
          description: Despesa atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
    delete:
      operationId: deleteDespesa
      tags: [financeiro]
      summary: Remover despesa
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Despesa removida
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /financeiro/despesas/{id}/comprovante:
    post:
      operationId: uploadComprovante
      tags: [financeiro]
      summary: Anexar comprovante URL a uma despesa
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comprovante:
                  type: string
              required: [comprovante]
      responses:
        "200":
          description: Comprovante salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
  /financeiro/despesas-recorrentes:
    get:
      operationId: listDespesasRecorrentes
      tags: [financeiro]
      summary: Listar despesas recorrentes
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Lista de despesas recorrentes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DespesaRecorrente"
    post:
      operationId: criarDespesaRecorrente
      tags: [financeiro]
      summary: Criar despesa recorrente
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarDespesaRecorrenteInput"
      responses:
        "200":
          description: Despesa recorrente criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DespesaRecorrente"
  /financeiro/despesas-recorrentes/{id}:
    delete:
      operationId: deleteDespesaRecorrente
      tags: [financeiro]
      summary: Remover despesa recorrente
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Despesa recorrente removida
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /financeiro/gerar-recorrentes:
    post:
      operationId: gerarRecorrentes
      tags: [financeiro]
      summary: Gerar despesas recorrentes para o mês
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                mes:
                  type: number
                ano:
                  type: number
      responses:
        "200":
          description: Despesas geradas
          content:
            application/json:
              schema:
                type: object
                properties:
                  geradas:
                    type: number
                required: [geradas]
  /financeiro/comissoes:
    get:
      operationId: listComissoes
      tags: [financeiro]
      summary: Listar comissões do mês
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: mes
          in: query
          schema:
            type: number
        - name: ano
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Lista de comissões
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comissao"
  /financeiro/comissoes-config:
    get:
      operationId: getComissoesConfig
      tags: [financeiro]
      summary: Obter configuração de comissões
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Configuração atual
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComissaoConfig"
    post:
      operationId: saveComissaoConfig
      tags: [financeiro]
      summary: Salvar configuração de comissão por vendedor
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vendedor:
                  type: string
                percentual:
                  type: number
              required: [vendedor, percentual]
      responses:
        "200":
          description: Configuração salva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComissaoConfig"
  /financeiro/dre:
    get:
      operationId: getDRE
      tags: [financeiro]
      summary: DRE (Demonstrativo de Resultado) do mês
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - name: mes
          in: query
          schema:
            type: number
        - name: ano
          in: query
          schema:
            type: number
      responses:
        "200":
          description: DRE calculado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DREData"
  /financeiro/resumo-diario:
    get:
      operationId: getResumoDiario
      tags: [financeiro]
      summary: Resumo financeiro do dia corrente
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Resumo do dia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumoDiario"
  /financeiro/metas:
    get:
      operationId: listMetas
      tags: [financeiro]
      summary: Listar metas do mês
      parameters:
        - name: mes
          in: query
          schema:
            type: number
        - name: ano
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Metas do mês
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meta"
    post:
      operationId: saveMeta
      tags: [financeiro]
      summary: Salvar meta mensal
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveMetaInput"
      responses:
        "200":
          description: Meta salva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
  /financeiro/alertas:
    get:
      operationId: getAlertas
      tags: [financeiro]
      summary: Alertas financeiros ativos
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alerta"
  /financeiro/evolucao:
    get:
      operationId: getEvolucao
      tags: [financeiro]
      summary: Evolução de vendas e despesas dos últimos meses
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Dados de evolução
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvolucaoMes"

  # ── Chat (SSE) ────────────────────────────────────────────────────────────
...
components:
  parameters:
    SessionToken:
      name: x-session-token
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    Despesa:
      type: object
      properties:
        id:
          type: number
        lojaId:
          type: number
        valor:
          type: string
        categoria:
          type: string
        descricao:
          type: string
          nullable: true
        comprovante:
          type: string
          nullable: true
        recorrente:
          type: boolean
        recorrenteId:
          type: number
          nullable: true
        confirmada:
          type: boolean
        data:
          type: string
        criadoEm:
          type: string
      required: [id, valor, categoria, recorrente, confirmada, data]
    CriarDespesaInput:
      type: object
      properties:
        valor:
          type: number
        categoria:
          type: string
        descricao:
          type: string
        comprovante:
          type: string
        recorrente:
          type: boolean
        data:
          type: string
      required: [valor, categoria]
    DespesaRecorrente:
      type: object
      properties:
        id:
          type: number
        valor:
          type: string
        categoria:
          type: string
        descricao:
          type: string
          nullable: true
        ativo:
          type: boolean
        diaVencimento:
          type: number
        criadoEm:
          type: string
      required: [id, valor, categoria, ativo, diaVencimento]
    CriarDespesaRecorrenteInput:
      type: object
      properties:
        valor:
          type: number
        categoria:
          type: string
        descricao:
          type: string
        diaVencimento:
          type: number
      required: [valor, categoria]
    Comissao:
      type: object
      properties:
        vendedor:
          type: string
        totalVendas:
          type: number
        percentual:
          type: number
        valorComissao:
          type: number
        quantidadeVendas:
          type: number
      required: [vendedor, totalVendas, percentual, valorComissao, quantidadeVendas]
    ComissaoConfig:
      type: object
      properties:
        id:
          type: number
        vendedor:
          type: string
        percentual:
          type: number
        criadoEm:
          type: string
      required: [id, vendedor, percentual]
    DREData:
      type: object
      properties:
        mes:
          type: number
        ano:
          type: number
        receitaBruta:
          type: number
        despesaTotal:
          type: number
        comissoesTotal:
          type: number
        lucroLiquido:
          type: number
        margemLiquida:
          type: number
        despesasPorCategoria:
          type: array
          items:
            type: object
            properties:
              categoria:
                type: string
              total:
                type: number
            required: [categoria, total]
      required: [mes, ano, receitaBruta, despesaTotal, lucroLiquido]
    ResumoDiario:
      type: object
      properties:
        vendas:
          type: number
        orcamentos:
          type: number
        faturamento:
          type: number
        despesas:
          type: number
      required: [vendas, orcamentos, faturamento, despesas]
    Meta:
      type: object
      properties:
        id:
          type: number
        mes:
          type: number
        ano:
          type: number
        metaFaturamento:
          type: number
        metaVendas:
          type: number
        criadoEm:
          type: string
      required: [id, mes, ano]
    SaveMetaInput:
      type: object
      properties:
        mes:
          type: number
        ano:
          type: number
        metaFaturamento:
          type: number
        metaVendas:
          type: number
      required: [mes, ano]
    Alerta:
      type: object
      properties:
        tipo:
          type: string
        mensagem:
          type: string
        severidade:
          type: string
          enum: [info, aviso, critico]
      required: [tipo, mensagem, severidade]
    EvolucaoMes:
      type: object
      properties:
        mes:
          type: string
        faturamento:
          type: number
        despesas:
          type: number
        lucro:
          type: number
      required: [mes, faturamento, despesas, lucro]
...