# lib/api-zod/src/generated/types/comissaoConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Castor Orçamento API
 * OpenAPI spec version: 0.1.0
 */

export interface ComissaoConfig {
  id: number;
  vendedor: string;
  percentual: number;
  criadoEm?: string;
}